pyproject.toml
[project]
name = "code-organizer"
version = "0.1.0"
description = "A comprehensive tool to scan, analyze, and organize scattered source code across your entire computer"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "rich>=13.0.0",
    "click>=8.1.0",
    "gitpython>=3.1.40",
    "PyGithub>=2.1.1",
    "python-gitlab>=4.4.0",
    "jinja2>=3.1.2",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "pandas>=2.1.0",
    "plotly>=5.17.0",
    "requests>=2.31.0",
    "tqdm>=4.66.0",
    "colorama>=0.4.6",
    "tabulate>=0.9.0",
    "questionary>=2.0.0",
    "networkx>=3.2.0",
]

[project.scripts]
code-organizer = "code_organizer.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
